### 2026-02-11 #9953 @editorialbot
[PRE REVIEW]: Musubi: Octree based Lattice-Boltzmann solver for multi-physics
**Submitting author:** @haraldkl (<a href="http://orcid.org/0000-0002-6054-5681">Harald Klimach</a>)
**Repository:** https://github.com/apes-suite/musubi
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.0.0
