### 2026-02-12 28db224
JOSS: Add newly requested scope sections
