+++ paper.bib
@@ -62,6 +62,20 @@
 	Year = {2013}
 }
 
+@article{Zimny:2013,
+	author = {Zimny, Simon and Chopard, Bastien and Malaspinas, Orestis and Lorenz, Eric and Jain, Kartik and Roller, Sabine and Bernsdorf, J{\"o}rg},
+	doi = {10.1016/j.procs.2013.05.266},
+	issn = {1877-0509},
+	journal = {Procedia Computer Science},
+	keywords = {Multiscale Simulation, Lattice Boltzmann Method, Scale Separation Map, Thrombus, Intracranial Aneurysms},
+	note = {2013 International Conference on Computational Science},
+	pages = {1006--1015},
+	title = {A Multiscale Approach for the Coupled Simulation of Blood Flow and Thrombus Formation in Intracranial Aneurysms},
+	volume = {18},
+	year = {2013},
+}
+
+
 @inproceedings{Klimach:2014,
 	address = {Munich, Germany},
 	doi = {10.3233/978-1-61499-381-0-703},
@@ -96,11 +110,50 @@
 	Year = {2015}
 }
 
+@inproceedings{Qi:2015,
+	author = {Qi, Jiaxing and Hasert, Manuel and Klimach, Harald and Roller, Sabine},
+	booktitle = {Sustained Simulation Performance 2015},
+	doi = {10.1007/978-3-319-20340-9_16},
+	editor = {Resch, Michael and Bez, Wolfgang and Focht, Erich and Kobayashi, Hiroaki and Qi, Jiaxing and Roller, Sabine},
+	pages = {195--204},
+	publisher = {Springer International Publishing},
+	title = {Aeroacoustic Simulation of Flow Through Porous Media Based on Lattice Boltzmann Method},
+	year = {2015}
+}
+
+
+@article{Johannink:2015,
+title = {Predictive pressure drop models for membrane channels with non-woven and woven spacers},
+journal = {Desalination},
+volume = {376},
+pages = {41-54},
+year = {2015},
+issn = {0011-9164},
+doi = {10.1016/j.desal.2015.07.024},
+url = {https://www.sciencedirect.com/science/article/pii/S0011916415300321},
+author = {Matthias Johannink and Kannan Masilamani and Adel Mhamdi and Sabine Roller and Wolfgang Marquardt},
+keywords = {Membrane processes, Spacer, Pressure drop, Modeling},
+}
+
+@article{JainCaF:2016,
+	annote = {Journal Papers},
+	author = {Jain, Kartik and Roller, Sabine and Mardal, Kent-Andr{\'e}},
+	doi = {doi:10.1016/j.compfluid.2015.12.011},
+	journal = {Computers \& Fluids},
+	month = {mar},
+	pages = {36--46},
+	publisher = {Pergamon},
+	title = {Transitional flow in intracranial aneurysms--a space and time refinement study below the Kolmogorov scales using Lattice Boltzmann Method},
+	volume = {127},
+	year = {2016}
+}
+
 @phdthesis{Jain:2016,
 	Annote = {Supervisor: Roller, Sabine},
 	Author = {Jain, Kartik},
 	School = {Universit{\"a}t Siegen},
 	Title = {Transition to Turbulence in Physiological Flows: Direct Numerical Simulation of Hemodynamics in Intracranial Aneurysms and Cerebrospinal Fluid Hydrodynamics in the Spinal Canal},
+	isbn = {978-3-936533-83-5},
 	Year = {2016}
 }
 
@@ -114,6 +167,22 @@
 	volume = {27},
 	year = {2016}
 }
+
+@article{Masilamani:2017,
+	annote = {Journal Papers},
[…]
