### 2026-02-13 61869e5
Add DOI fields to bibliography entries
