+++ paper.bib
@@ -162,6 +162,7 @@ keywords = {Membrane processes, Spacer, Pressure drop, Modeling},
 	booktitle = {Parallel Computing: On the Road to Exascale},
 	editor = {Gerhard R. Joubert and Hugh Leather and Mark Parsons and Frans Peters and Mark Sawyer},
 	pages = {807--816},
+    doi = {10.3233/978-1-61499-621-7-807},
 	series = {Advances in Parallel Computing},
 	title = {Performance Evaluation of the LBM Solver Musubi on Various HPC Architectures},
 	volume = {27},
@@ -189,6 +190,7 @@ keywords = {Membrane processes, Spacer, Pressure drop, Modeling},
 	Author = {Qi, Jiaxing},
 	School = {RWTH Aachen University},
 	Title = {Efficient Lattice Boltzmann Simulations on Large Scale High Performance Computing Systems},
+    doi = {10.18154/RWTH-2017-04121},
 	Year = {2017}
 }
 
@@ -197,6 +199,7 @@ keywords = {Membrane processes, Spacer, Pressure drop, Modeling},
 	Author = {Masilamani, Kannan},
 	School = {RWTH Aachen University},
 	Title = {Framework for Coupled Simulation of Electrodialysis Processes},
+    doi = {10.18154/RWTH-2021-02437},
 	Year = {2020}
 }
 
