### 2026-02-13 3da5d3d
joss: Fix typo in Acknowledgements section header
