### 2026-02-13 ba55123
Merge Musubi section into Design
* Moved the lattice Boltzmann section further up.

* Removed redundant sections and improved clarity on software design and research impact.
