### 2026-02-13 #9953 @haraldkl
Dear @arfon, thanks a lot for picking up on our submission. I've adapted the paper to hopefully meet the new layout ([ba55123](https://github.com/apes-suite/musubi/tree/ba55123f0f0c8d623b9eb9337a84c44b1e1ac044)), and address all the points with respect to the scope.

I would like to point out, that the repository structure is a little bit involved as it makes use of git submodules. The actual source code of the project resides in https://github.com/apes-suite/musubi-source. The project started out in Mercurial with subrepositories and was hosted on OSDN. After that hoster became unavailable, we decided to move on to git and github, but truncated the history due to the complications with subrepositories. Now testing, documentation and issue tracking all are hosted here on github.
