+++ src/mus_auxFieldVar_module.f90
@@ -514,8 +514,8 @@ contains
     ! vectorization
     nChunks = ceiling(real(nSolve, kind=rk) / real(vlen, kind=rk))
 
-    !$omp parallel do                                                                            &
-    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel ) &
+    !$omp parallel do                                                                          &
+    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel )
     do iChunks = 1, nChunks
       ! calculate the end  number of iElem loop
       nChunkElems = min(vlen, nSolve - ((iChunks - 1) * vlen))
@@ -606,8 +606,8 @@ contains
     ! vectorization
     nChunks = ceiling(real(nSolve, kind=rk) / real(vlen, kind=rk))
 
-    !$omp parallel do                                                                            &
-    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel ) &
+    !$omp parallel do                                                                          &
+    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel )
     do iChunks = 1, nChunks
       ! calculate the end  number of iElem loop
       nChunkElems = min(vlen, nSolve - ((iChunks - 1) * vlen))
@@ -699,8 +699,8 @@ contains
     ! vectorization
     nChunks = ceiling(real(nSolve, kind=rk) / real(vlen, kind=rk))
 
-    !$omp parallel do                                                                            &
-    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel ) &
+    !$omp parallel do                                                                          &
+    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel )
     do iChunks = 1, nChunks
       ! calculate the end  number of iElem loop
       nChunkElems = min(vlen, nSolve - ((iChunks - 1) * vlen))
@@ -805,8 +805,8 @@ contains
     ! vectorization
     nChunks = ceiling(real(nSolve, kind=rk) / real(vlen, kind=rk))
 
-    !$omp parallel do                                                                            &
-    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel ) &
+    !$omp parallel do                                                                          &
+    !$omp   & private( rho, pdf, elemPos, elemOff, nChunkElems, iChunks, low_bound, iElem, vel )
     do iChunks = 1, nChunks
       ! calculate the end  number of iElem loop
       nChunkElems = min(vlen, nSolve - ((iChunks - 1) * vlen))
