+++ src/mus_construction_module.f90
@@ -11,6 +11,7 @@
 ! Copyright (c) 2016 Peter Vitt <peter.vitt2@uni-siegen.de>
 ! Copyright (c) 2016-2018 Raphael Haupt <raphael.haupt@uni-siegen.de>
 ! Copyright (c) 2021 Gregorio Gerardo Spinelli <gregoriogerardo.spinelli@dlr.de>
+! Copyright (c) 2025 Mengyu Wang <m.wang-2@utwente.nl>
 !
 ! Redistribution and use in source and binary forms, with or without
 ! modification, are permitted provided that the following conditions are met:
@@ -257,6 +258,7 @@ contains
     type( logical_array_type ), allocatable :: haloRequired(:)
     integer :: symmetricBCs(geometry%boundary%nBCtypes)
     integer :: nSymBCs
+    logical :: needAuxHaloCommDefault
     ! --------------------------------------------------------------------------
     minLevel = geometry%tree%global%minLevel
     maxLevel = geometry%tree%global%maxLevel
@@ -276,14 +278,14 @@ contains
     !    boundaries
     ! 4. allocate the BC lists in globbc
     ! THIS ROUTINE DOES NOT MODIFY THE BC LISTS IN THE FIELDS!!!
-    call build_BClists( globBC    = scheme%globBC,              &
-      &                 tree      = geometry%tree,              &
-      &                 bc_prop   = geometry%boundary,          &
-      &                 minLevel  = minLevel,                   &
-      &                 maxLevel  = maxLevel,                   &
-      &                 layout    = scheme%layout,              &
-      &                 field     = scheme%field(:),            &
-      &                 comm      = params%general%proc%comm    )
+    call build_BClists( globBC    = scheme%globBC,           &
+      &                 tree      = geometry%tree,           &
+      &                 bc_prop   = geometry%boundary,       &
+      &                 minLevel  = minLevel,                &
+      &                 maxLevel  = maxLevel,                &
+      &                 layout    = scheme%layout,           &
+      &                 field     = scheme%field(:),         &
+      &                 comm      = params%general%proc%comm )
 
     ! Treat the boundary lists: in addition to the normal stencil treatment
     ! one has to find all those elements along the element list in the stencil
@@ -311,9 +313,9 @@ contains
     enddo ! iField
 
     ! Build and append IBM stencils to stencil array
-    call mus_build_IBMStencils( globIBM    = geometry%globIBM,         &
-      &                         layout     = scheme%layout,            &
-      &                         grwStencil = scheme%layout%grwStencil  )
+    call mus_build_IBMStencils( globIBM    = geometry%globIBM,        &
+      &                         layout     = scheme%layout,           &
+      &                         grwStencil = scheme%layout%grwStencil )
 
     ! finalize scheme layout
     ! copy growing array of stencil to allocatable array of stencil
@@ -332,7 +334,7 @@ contains
     !    directly
     ! THIS MEANS THAT FOR ALL ELEMENTS (INCL. BOUNDARY ELEMENTS) NEIGHBORS
     ! ARE STORED BUT THEY DO NOT NEED TO EXIST IN THE MESH AT THIS POINT!!!
-    call tem_init_elemLevels( me       = scheme%levelDesc,   &
+    call tem_init_elemLevels( me       = scheme%levelDesc,        &
       &                       boundary = geometry%boundary,       &
       &                       tree     = geometry%tree,           &
       &                       stencils = scheme%layout%stencil(:) )
@@ -379,11 +381,11 @@ contains
     write(dbgUnit(6),*) 'before horizontal dep'
     write(logUnit(6),*) 'before tem_build_horizontalDependencies'
     do iStencil = 1, scheme%layout%nStencils
[…]
