+++ src/mus_auxField_module.f90
@@ -1,4 +1,5 @@
 ! Copyright (c) 2019-2020, 2022 Kannan Masilamani <kannan.masilamani@uni-siegen.de>
+! Copyright (c) 2025 Mengyu Wang <m.wang-2@utwente.nl>
 !
 ! Redistribution and use in source and binary forms, with or without
 ! modification, are permitted provided that the following conditions are met:
@@ -39,6 +40,8 @@ module mus_auxField_module
     &                           tem_comm_init
   use tem_construction_module, only: tem_levelDesc_type
   use tem_general_module,      only: tem_general_type
+  use tem_tracking_module,     only: tem_tracking_type
+  use tem_logging_module,      only: logUnit
 
   use mus_derVarPos_module,          only: mus_derVarPos_type
   use mus_scheme_header_module,      only: mus_scheme_header_type
@@ -60,6 +63,7 @@ module mus_auxField_module
   public :: mus_calcAuxFieldAndExchange
   public :: mus_intpAuxFieldCoarserAndExchange
   public :: mus_intpAuxFieldFinerAndExchange
+  public :: mus_auxField_configure_from_tracking
 
   !> Contains auxiliary field variable values per level and communication
   !! buffers
@@ -85,6 +89,8 @@ module mus_auxField_module
     type( tem_communication_type ) :: recvBufferFromCoarser
     !> My halos which are ghostFromFiner on remote processes
     type( tem_communication_type ) :: recvBufferFromFiner
+    !> Controls whether auxField halo exchange is required on this level.
+    logical :: needHaloComm = .false.
   end type mus_auxFieldVar_type
 
   abstract interface
@@ -127,12 +133,13 @@ module mus_auxField_module
   end interface
 
 
-
 contains
+
 
   ! ************************************************************************* !
   !> This routine initialize auxField var val array and communication buffers
-  subroutine mus_init_auxFieldArrays(me, levelDesc, pattern, nSize, nAuxScalars)
+  subroutine mus_init_auxFieldArrays(me, levelDesc, pattern, nSize, &
+    &                                nAuxScalars, needHaloComm      )
     ! --------------------------------------------------------------------- !
     !> Auxiliary field variable
     type(mus_auxFieldVar_type), intent(out) :: me
@@ -144,8 +151,14 @@ contains
     integer, intent(in) :: nSize
     !> Number of scalars in auxiliary variables
     integer, intent(in) :: nAuxScalars
+    !> Default for halo communication flag
+    logical, intent(in), optional :: needHaloComm
     ! --------------------------------------------------------------------- !
     ! --------------------------------------------------------------------- !
+    if (present(needHaloComm)) then
+      me%needHaloComm = needHaloComm
+    end if
+
     allocate(me%val(nSize * nAuxScalars))
     me%val(:) = -1000000.0_rk
 
@@ -223,7 +236,8 @@ contains
   !! PDF initialized by IC
   subroutine mus_initAuxFieldFluidAndExchange(auxField, state, neigh, nElems,  &
     &                                         nSize, nFields, stencil, varSys, &
-    &                                         derVarPos, iLevel, general, quantities)
+    &                                         derVarPos, iLevel, general,      &
+    &                                         quantities)
     !---------------------------------------------------------------------------
     !> auxilary field array
     type(mus_auxFieldVar_type), intent(inout) :: auxField
@@ -269,12 +283,14 @@ contains
     ! communicate velocity field. Requires for tubulence to compute ShearRate
     ! from velocity gradient.
     ! exchange velocity halo on current level
-    call general%commpattern%exchange_real(   &
-      &  send         = auxField%sendBuffer,  &
-      &  recv         = auxField%recvBuffer, &
[…]
