### 2026-04-14 #9953 @espottesmith
Hey @haraldkl, my name is Evan, and I'll be the managing Topic Editor for your JOSS submission.

To start us off, could you suggest some potential reviewers? You could look at the authors of the manuscripts listed above, or you could take a look at our [JOSS database](reviewers.joss.theoj.org/reviewers).
