### 2026-04-22 #9953 @espottesmith
Oh, sorry. Seems I messed up the link. It's https://reviewers.joss.theoj.org/reviewers
