### 2026-04-22 #9953 @espottesmith
@Yazhuo-Liu @esclapez @jayten @zbeekman @jamiebramwell

Can any of y'all contribute your expertise to review this JOSS submission?
