### 2026-08-16 #10561 @espottesmith
@TJFord @smillerc @ajayrawat @shurikkuzmin @hsalehipour @gouarin

Can any of you contribute reviews to this submission for JOSS?
