### 2026-08-25 #10561 @espottesmith
@desjardi @sorush-khajepor @AminZar69 @rubenzorrilla

Can any of you review this submission for JOSS? If you cannot (due to lack of time, lack of experience, etc.), it would be very helpful if you could suggest alternate reviewers!
