### 2026-08-28 #10561 @ajayrawat
I am able to install the software using the waf script inside my wsl with Ubuntu 24.04.4 LTS. It is a good start:
```
 ./bin/waf configure build
  tests that fail 0/62
Utests successful!
'build' finished successfully (2m1.726s)
```
