+++ src/mus_derQuan_module.f90
@@ -235,6 +235,8 @@ module mus_derQuan_module
   public :: derive_HRRCorrection_d3q27
   public :: derive_brinkmanForce
   public :: derive_brinkmanForce_TRT
+  public :: derive_brinkmanForce_2ndOrd_BGK
+  public :: derive_brinkmanForce_2ndOrd_TRT
 
   ! Apply source add source term to state
   public :: applySrc_absorbLayer
@@ -242,6 +244,7 @@ module mus_derQuan_module
   public :: applySrc_absorbLayerDyn
   public :: applySrc_absorbLayerDyn_MRT
   public :: applySrc_force
+  public :: applySrc_force_TRT
   public :: applySrc_force_GNS
   public :: applySrc_force_MRT
   public :: applySrc_force_MRT_d2q9
@@ -255,6 +258,8 @@ module mus_derQuan_module
   public :: applySrc_force1stOrd
   public :: applySrc_brinkmanForce
   public :: applySrc_brinkmanForce_TRT
+  public :: applySrc_brinkmanForce_2ndOrd_BGK
+  public :: applySrc_brinkmanForce_2ndOrd_TRT
 
 contains
 
@@ -2422,7 +2427,6 @@ contains
       do iDir = 1, QQ
         ucx = dot_product( scheme%layout%fStencil%cxDirRK(:, iDir), &
           &                velocity )
-
         res( (iElem-1) * fun%nComponents + iDir ) = (omegaKine / 2.0_rk - 1.0_rk)     &
           &                                        * scheme%layout%weight(iDir)       &
           &                                        * cs2inv * bCoeffField(iElem) * ucx * rho0
@@ -2538,7 +2542,6 @@ contains
       do iDir = 1, QQ
         ucx = dot_product( scheme%layout%fStencil%cxDirRK(:, iDir), &
           &                velocity )
-
         res( (iElem - 1) * fun%nComponents + iDir ) = (omegaMinus / 2.0_rk - 1.0_rk)    &
           &                                          * scheme%layout%weight(iDir)       &
           &                                          * cs2inv * bCoeffField(iElem) * ucx * rho0
@@ -2548,6 +2551,243 @@ contains
     end do !iElem
 
   end subroutine derive_brinkmanForce_TRT
+! ****************************************************************************** !
+
+! ****************************************************************************** !
+   !> Derive the Brinkman force variable defined as a source term.
+   !!
+   !! It multiplies the Brinkman coefficient variable as defined by the user
+   !! with the negative velocity field to obtain the force term.
+   !! This force term is then converted to a state value that is to be added to the
+   !! state.
+   !!
+   !! Reference:
+   !! 1) Zhaoli Guo and T. S. Zhao. “Lattice Boltzmann Model for Incompressible
+   !!    Flows through Porous Media”. In: Physical Review E 66.3 (2002), p. 036304.
+   !!    doi: 10.1103/PhysRevE.66.036304.
+   !! 2) Irina Ginzburg. “Consistent lattice Boltzmann schemes for the Brinkman
+   !!    model of porous flow and infinite Chapman-Enskog expansion”. In: Phys.
+   !!    Rev. E 77 (6 June 2008), p. 066704. doi: 10.1103/PhysRevE.77.066704.
+  recursive subroutine derive_brinkmanForce_2ndOrd_BGK(fun, varsys, elempos, time, &
+    &                                                  tree, nElems, nDofs, res    )
+    ! -------------------------------------------------------------------- !
+    !> Description of the method to obtain the variables, here some preset
+    !! values might be stored, like the space time function to use or the
+    !! required variables.
+    class(tem_varSys_op_type), intent(in) :: fun
+
+    !> The variable system to obtain the variable from.
[…]
