### 2026-09-20 d7d1794
Add references in summary
This commit addresses the following reviewer comments:

1. APES is abbreviated, but it full form should be provided line 8.
   -> Adaptable Poly-Engineering Simulator is now introduced in full
2. Provide ref for 1st paragraph
   -> Not quite sure, but the summary now has a reference to the
      APES-Suite github organization and a reference for Space-Filling
      curves
3. It is mentioned that program is written in Fortran 2003 standard,
   author should also mention something about MPI.
   -> The use of MPI with MPI-3 features is now mentioned and MPI-3
   referenc
[…]
