+++ paper.bib
@@ -11,6 +11,62 @@
 }
 
 
+@Book{Krueger:2016,
+  author = {Kr{\"u}ger, Timm
+            and Kusumaatmaja, Halim
+            and Kuzmin, Alexandr
+            and Shardt, Orest
+            and Silva, Goncalo
+            and Viggen, Erlend Magnus},
+  publisher = {Springer Cham},
+  title = {The Lattice Boltzmann Method},
+  year = {2016},
+  edition = {First edition},
+  isbn = {9783319446479},
+  pagetotal = {694},
+  doi = {10.1007/978-3-319-44649},
+}
+
+
+@Book{Sagan:1994,
+  author = {Hans Sagan},
+  publisher = {Springer New York, NY},
+  title = {Space-Filling Curves},
+  year = {1994},
+  edition = {First edition},
+  isbn = {9780387942650},
+  pagetotal = {194},
+  doi = {10.1007/978-1-4612-0871-6},
+}
+
+
+@manual{MPI3:2012,
+  author = {{MPI Forum}},
+  title = {{MPI}: A Message-Passing Interface Standard, Version 3.0},
+  year = {2012},
+  month = {sep},
+  url = {https://www.mpi-forum.org/docs/mpi-3.0/mpi30-report.pdf},
+  note = {Last accessed: 2026-09-20},
+}
+
+
+@misc{APES,
+  author = {Klimach, Harald
+            and Masilamani, Kannan
+            and Jain, Kartik
+            and Vlogman, Tristan
+            and Wang, Mengyu
+            and Gericke-Schuster, Jana
+	    and Ebrahimi Pour, Neda
+            and Vitt, Peter},
+  title = {Adaptable Poly-Engineering Simulator (APES)},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub organization},
+  url = {https://github.com/apes-suite},
+}
+
+
 @article{Klimach:2012vi,
     author = {Klimach, Harald and Hasert, Manuel and Zudrop, Jens and Roller, Sabine},
     title = {{Distributed Octree Mesh Infrastructure for Flow Simulations}},
@@ -123,16 +179,16 @@
 
 
 @article{Johannink:2015,
-title = {Predictive pressure drop models for membrane channels with non-woven and woven spacers},
-journal = {Desalination},
-volume = {376},
-pages = {41-54},
-year = {2015},
-issn = {0011-9164},
-doi = {10.1016/j.desal.2015.07.024},
-url = {https://www.sciencedirect.com/science/article/pii/S0011916415300321},
-author = {Matthias Johannink and Kannan Masilamani and Adel Mhamdi and Sabine Roller and Wolfgang Marquardt},
-keywords = {Membrane processes, Spacer, Pressure drop, Modeling},
+  title = {Predictive pressure drop models for membrane channels with non-woven and woven spacers},
+  journal = {Desalination},
+  volume = {376},
+  pages = {41-54},
+  year = {2015},
+  issn = {0011-9164},
+  doi = {10.1016/j.desal.2015.07.024},
+  url = {https://www.sciencedirect.com/science/article/pii/S0011916415300321},
+  author = {Matthias Johannink and Kannan Masilamani and Adel Mhamdi and Sabine Roller and Wolfgang Marquardt},
+  keywords = {Membrane processes, Spacer, Pressure drop, Modeling},
 }
 
 @article{JainCaF:2016,
